+++ sdk/metric/internal/aggregate/lastvalue.go
@@ -21,13 +21,105 @@ type lastValuePoint[N int64 | float64] struct {
 	dropExemplars bool
 }
 
-// lastValueMap summarizes a set of measurements as the last one made.
-type lastValueMap[N int64 | float64] struct {
-	newRes func(attribute.Set) FilteredExemplarReservoir[N]
-	values limitedSyncMap[*lastValuePoint[N]]
+func newDeltaLastValue[N int64 | float64](
+	limit int,
+	r func(attribute.Set) FilteredExemplarReservoir[N],
+) *deltaLastValue[N] {
+	return &deltaLastValue[N]{
+		newRes: r,
+		start:  now(),
+		vals:   newHotColdMap[*lastValuePoint[N]](limit),
+	}
 }
 
-func (s *lastValueMap[N]) measure(
+// deltaLastValue summarizes a set of measurements as the last one made.
+type deltaLastValue[N int64 | float64] struct {
+	newRes func(attribute.Set) FilteredExemplarReservoir[N]
+	start  time.Time
+	vals   *hotColdMap[*lastValuePoint[N]]
+}
+
+func (s *deltaLastValue[N]) measure(
+	ctx context.Context,
+	value N,
+	fltrAttr attribute.Set,
+	droppedAttr []attribute.KeyValue,
+) {
+	s.vals.WriteUnbound(fltrAttr, func(attr attribute.Set) *lastValuePoint[N] {
+		r := s.newRes(attr)
+		_, isDrop := r.(*dropRes[N])
+		return &lastValuePoint[N]{
+			res:           r,
+			attrs:         attr,
+			startTime:     now(),
+			dropExemplars: isDrop,
+		}
+	}, func(lv *lastValuePoint[N]) {
+		lv.value.Store(value)
+		if !lv.dropExemplars {
+			lv.res.Offer(ctx, value, droppedAttr)
+		}
+	})
+}
+
+func (s *deltaLastValue[N]) collect(
+	dest *metricdata.Aggregation, //nolint:gocritic // The pointer is needed for the ComputeAggregation interface
+) int {
+	t := now()
+	n := s.copyAndClearDpts(dest, t)
+	// Update start time for delta temporality.
+	s.start = t
+	return n
+}
+
+// copyAndClearDpts copies the lastValuePoints held by s into dest. The number of lastValuePoints
+// copied is returned.
+func (s *deltaLastValue[N]) copyAndClearDpts(
+	dest *metricdata.Aggregation, //nolint:gocritic // The pointer is needed for the ComputeAggregation interface
+	t time.Time,
+) int {
+	// Ignore if dest is not a metricdata.Gauge. The chance for memory reuse of
+	// the DataPoints is missed (better luck next time).
+	gData, _ := (*dest).(metricdata.Gauge[N])
+	readIdx := s.vals.SwapHotAndWait()
+	n := s.vals.Len(readIdx)
+	dPts := reset(gData.DataPoints, 0, n)
+
+	s.vals.Collect(readIdx, func(*lastValuePoint[N]) bool { return false }, func(_ any, v *lastValuePoint[N]) bool {
+		dp := metricdata.DataPoint[N]{
+			Attributes: v.attrs,
+			StartTime:  s.start,
+			Time:       t,
+			Value:      v.value.Load(),
+		}
+		collectExemplars[N](&dp.Exemplars, v.res.Collect)
+		dPts = append(dPts, dp)
+		return true
+	})
+	gData.DataPoints = dPts
+	*dest = gData
+	return len(dPts)
+}
+
+// cumulativeLastValue summarizes a set of measurements as the last one made.
+type cumulativeLastValue[N int64 | float64] struct {
+	newRes func(attribute.Set) FilteredExemplarReservoir[N]
+	values limitedSyncMap[*lastValuePoint[N]]
+	start  time.Time
+}
+
+func newCumulativeLastValue[N int64 | float64](
+	limit int,
+	r func(attribute.Set) FilteredExemplarReservoir[N],
+) *cumulativeLastValue[N] {
+	return &cumulativeLastValue[N]{
+		newRes: r,
+		values: limitedSyncMap[*lastValuePoint[N]]{state: &cardinalityState{limit: limit}},
+		start:  now(),
+	}
+}
+
+func (s *cumulativeLastValue[N]) measure(
 	ctx context.Context,
 	value N,
 	fltrAttr attribute.Set,
@@ -52,109 +144,6 @@ func (s *lastValueMap[N]) measure(
 	}
 }
 
-func newDeltaLastValue[N int64 | float64](
-	limit int,
-	r func(attribute.Set) FilteredExemplarReservoir[N],
-) *deltaLastValue[N] {
-	return &deltaLastValue[N]{
-		newRes: r,
-		start:  now(),
-		hotColdValMap: [2]lastValueMap[N]{
-			{
-				newRes: r,
-				values: limitedSyncMap[*lastValuePoint[N]]{aggLimit: limit},
-			},
-			{
-				newRes: r,
-				values: limitedSyncMap[*lastValuePoint[N]]{aggLimit: limit},
-			},
-		},
-	}
-}
-
-// deltaLastValue summarizes a set of measurements as the last one made.
-type deltaLastValue[N int64 | float64] struct {
-	newRes func(attribute.Set) FilteredExemplarReservoir[N]
-	start  time.Time
-
-	hcwg          hotColdWaitGroup
-	hotColdValMap [2]lastValueMap[N]
-}
-
-func (s *deltaLastValue[N]) measure(
-	ctx context.Context,
-	value N,
-	fltrAttr attribute.Set,
-	droppedAttr []attribute.KeyValue,
-) {
-	hotIdx := s.hcwg.start()
-	defer s.hcwg.done(hotIdx)
-	s.hotColdValMap[hotIdx].measure(ctx, value, fltrAttr, droppedAttr)
-}
-
-func (s *deltaLastValue[N]) collect(
-	dest *metricdata.Aggregation, //nolint:gocritic // The pointer is needed for the ComputeAggregation interface
-) int {
-	t := now()
-	n := s.copyAndClearDpts(dest, t)
-	// Update start time for delta temporality.
-	s.start = t
-	return n
-}
-
-// copyAndClearDpts copies the lastValuePoints held by s into dest. The number of lastValuePoints
-// copied is returned.
-func (s *deltaLastValue[N]) copyAndClearDpts(
-	dest *metricdata.Aggregation, //nolint:gocritic // The pointer is needed for the ComputeAggregation interface
-	t time.Time,
-) int {
-	// Ignore if dest is not a metricdata.Gauge. The chance for memory reuse of
-	// the DataPoints is missed (better luck next time).
-	gData, _ := (*dest).(metricdata.Gauge[N])
-	// delta always clears values on collection
-	readIdx := s.hcwg.swapHotAndWait()
-	// The len will not change while we iterate over values, since we waited
-	// for all writes to finish to the cold values and len.
-	n := s.hotColdValMap[readIdx].values.Len()
-	dPts := reset(gData.DataPoints, n, n)
-
-	var i int
-	s.hotColdValMap[readIdx].values.Range(func(_, value any) bool {
-		v := value.(*lastValuePoint[N])
-		dPts[i].Attributes = v.attrs
-		dPts[i].StartTime = s.start
-		dPts[i].Time = t
-		dPts[i].Value = v.value.Load()
-		collectExemplars[N](&dPts[i].Exemplars, v.res.Collect)
-		i++
-		return true
-	})
-	gData.DataPoints = dPts
-	// Do not report stale values.
-	s.hotColdValMap[readIdx].values.Clear()
-	*dest = gData
-	return i
-}
-
-// cumulativeLastValue summarizes a set of measurements as the last one made.
-type cumulativeLastValue[N int64 | float64] struct {
-	lastValueMap[N]
-	start time.Time
-}
-
-func newCumulativeLastValue[N int64 | float64](
-	limit int,
-	r func(attribute.Set) FilteredExemplarReservoir[N],
-) *cumulativeLastValue[N] {
-	return &cumulativeLastValue[N]{
-		lastValueMap: lastValueMap[N]{
-			newRes: r,
-			values: limitedSyncMap[*lastValuePoint[N]]{aggLimit: limit},
-		},
-		start: now(),
-	}
-}
-
 func (s *cumulativeLastValue[N]) collect(
 	dest *metricdata.Aggregation, //nolint:gocritic // The pointer is needed for the ComputeAggregation interface
 ) int {
@@ -170,9 +159,7 @@ func (s *cumulativeLastValue[N]) collect(
 	perSeriesStartTimeEnabled := x.PerSeriesStartTimestamps.Enabled()
 
 	var i int
-	s.values.Range(func(_, value any) bool {
-		v := value.(*lastValuePoint[N])
-
+	s.values.Range(func(_ any, v *lastValuePoint[N]) bool {
 		startTime := s.start
 		if perSeriesStartTimeEnabled {
 			startTime = v.startTime